% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\docType{methods}
\name{dbConnect,OdbcDriver-method}
\alias{dbConnect,OdbcDriver-method}
\alias{dbConnect}
\title{Connect to a ODBC compatible database}
\usage{
\S4method{dbConnect}{OdbcDriver}(drv, dsn = NULL, ..., timezone = "UTC",
  driver = NULL, server = NULL, database = NULL, uid = NULL,
  pwd = NULL, .connection_string = NULL)
}
\arguments{
\item{drv}{an object that inherits from \linkS4class{DBIDriver},
or an existing \linkS4class{DBIConnection}
object (in order to clone an existing connection).}

\item{dsn}{The Data Source Name.}

\item{...}{Additional ODBC keywords, these will be joined with the other
arguments to form the final connection string.}

\item{timezone}{The Server time zone. Useful if the database has an internal
timezone that is \emph{not} 'UTC'. If the database is in your local timezone set
to \code{Sys.timezone()}. See \code{\link[=OlsonNames]{OlsonNames()}} for a complete list of available
timezones on your system.}

\item{driver}{The ODBC driver name.}

\item{server}{The server hostname.}

\item{database}{The database on the server.}

\item{uid}{The user identifer.}

\item{pwd}{The password to use.}

\item{.connection_string}{A complete connection string, useful if you are
copy pasting it from another source. If this argument is used any additional
arguments will be appended to this string.}

\item{encoding}{The Server text encoding. Used if the database has an
internal encoding that is \emph{not} \code{UTF-8}. If the database is in your local
encoding set to \code{""}. See \code{\link[=iconvlist]{iconvlist()}} for a complete list of available
encodings on your system.}
}
\description{
Connect to a ODBC compatible database
}
\details{
The connection string keywords are driver dependant. The parameters
documented here are common, but some drivers may not accept them. Please see
the specific driver documentation for allowed parameters,
\url{https://www.connectionstrings.com} is also a useful resource of example
connection strings for a variety of databases.
}
