% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\docType{methods}
\name{dbFetch,ODBConnectResult-method}
\alias{dbFetch,ODBConnectResult-method}
\title{Retrieve records from ODBConnect query}
\usage{
\S4method{dbFetch}{ODBConnectResult}(res, n = -1, ...)
}
\description{
Retrieve records from ODBConnect query
}

