% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R, R/db.R, R/driver-db2.R
\docType{class}
\name{DBI-classes}
\alias{DBI-classes}
\alias{Oracle-class}
\alias{Teradata-class}
\alias{HDB-class}
\alias{Hive-class}
\alias{Spark SQL-class}
\alias{DB2/AIX64-class}
\title{DBI classes}
\description{
DBI classes
}
\keyword{internal}
