% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R, R/Table.R, R/driver-db2.R,
%   R/driver-hana.R, R/driver-hive.R, R/driver-oracle.R, R/driver-terdata.R
\name{DBI-methods}
\alias{DBI-methods}
\alias{sqlData,OdbcConnection-method}
\alias{sqlCreateTable,DB2/AIX64-method}
\alias{sqlCreateTable,HDB-method}
\alias{dbQuoteString,Hive,character-method}
\alias{sqlCreateTable,Oracle-method}
\alias{sqlCreateTable,Teradata-method}
\title{Unimportant DBI methods}
\usage{
\S4method{sqlData}{OdbcConnection}(con, value, row.names = NA, ...)

\S4method{sqlCreateTable}{HDB}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)

\S4method{dbQuoteString}{Hive,character}(conn, x, ...)

\S4method{sqlCreateTable}{Oracle}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)

\S4method{sqlCreateTable}{Teradata}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)
}
\arguments{
\item{...}{Other parameters passed on to methods.}

\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}
}
\description{
Unimportant DBI methods
}
\keyword{internal}
