% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R, R/Connection.R, R/Result.R, R/Table.R
\docType{class}
\name{DBI}
\alias{DBI}
\alias{OdbconnectConnection-class}
\alias{OdbconnectDriver-class}
\alias{OdbconnectResult-class}
\alias{dbBegin,OdbconnectConnection-method}
\alias{dbBind,OdbconnectResult-method}
\alias{dbClearResult,OdbconnectResult-method}
\alias{dbColumnInfo,OdbconnectResult-method}
\alias{dbCommit,OdbconnectConnection-method}
\alias{dbConnect,OdbconnectDriver-method}
\alias{dbDataType,OdbconnectConnection,ANY-method}
\alias{dbDataType,OdbconnectDriver,ANY-method}
\alias{dbDataType,OdbconnectDriver,list-method}
\alias{dbDisconnect,OdbconnectConnection-method}
\alias{dbExistsTable,OdbconnectConnection,character-method}
\alias{dbFetch,OdbconnectResult-method}
\alias{dbGetInfo,OdbconnectConnection-method}
\alias{dbGetInfo,OdbconnectDriver-method}
\alias{dbGetInfo,OdbconnectResult-method}
\alias{dbGetRowCount,OdbconnectResult-method}
\alias{dbGetRowsAffected,OdbconnectResult-method}
\alias{dbGetStatement,OdbconnectResult-method}
\alias{dbHasCompleted,OdbconnectResult-method}
\alias{dbIsValid,OdbconnectConnection-method}
\alias{dbIsValid,OdbconnectDriver-method}
\alias{dbIsValid,OdbconnectResult-method}
\alias{dbListFields,OdbconnectConnection,character-method}
\alias{dbListTables,OdbconnectConnection-method}
\alias{dbQuoteIdentifier,OdbconnectConnection,character-method}
\alias{dbQuoteString,OdbconnectConnection,character-method}
\alias{dbReadTable,OdbconnectConnection,character-method}
\alias{dbRemoveTable,OdbconnectConnection,character-method}
\alias{dbRollback,OdbconnectConnection-method}
\alias{dbSendQuery,OdbconnectConnection,character-method}
\alias{dbSendStatement,OdbconnectConnection,character-method}
\alias{show,OdbconnectConnection-method}
\alias{show,OdbconnectDriver-method}
\alias{show,OdbconnectResult-method}
\title{DBI methods}
\usage{
\S4method{show}{OdbconnectDriver}(object)

\S4method{dbConnect}{OdbconnectDriver}(drv, dsn = NULL, ...)

\S4method{dbDataType}{OdbconnectDriver,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{OdbconnectDriver,list}(dbObj, obj, ...)

\S4method{dbIsValid}{OdbconnectDriver}(dbObj, ...)

\S4method{dbGetInfo}{OdbconnectDriver}(dbObj, ...)

\S4method{show}{OdbconnectConnection}(object)

\S4method{dbIsValid}{OdbconnectConnection}(dbObj, ...)

\S4method{dbDisconnect}{OdbconnectConnection}(conn, ...)

\S4method{dbSendQuery}{OdbconnectConnection,character}(conn, statement, ...)

\S4method{dbSendStatement}{OdbconnectConnection,character}(conn, statement, ...)

\S4method{dbDataType}{OdbconnectConnection,ANY}(dbObj, obj, ...)

\S4method{dbQuoteString}{OdbconnectConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{OdbconnectConnection,character}(conn, x, ...)

\S4method{dbListTables}{OdbconnectConnection}(conn, ...)

\S4method{dbExistsTable}{OdbconnectConnection,character}(conn, name, ...)

\S4method{dbListFields}{OdbconnectConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{OdbconnectConnection,character}(conn, name, ...)

\S4method{dbGetInfo}{OdbconnectConnection}(dbObj, ...)

\S4method{dbBegin}{OdbconnectConnection}(conn, ...)

\S4method{dbCommit}{OdbconnectConnection}(conn, ...)

\S4method{dbRollback}{OdbconnectConnection}(conn, ...)

\S4method{show}{OdbconnectResult}(object)

\S4method{dbClearResult}{OdbconnectResult}(res, ...)

\S4method{dbFetch}{OdbconnectResult}(res, n = -1, ..., row.names = NA)

\S4method{dbHasCompleted}{OdbconnectResult}(res, ...)

\S4method{dbGetInfo}{OdbconnectResult}(dbObj, ...)

\S4method{dbIsValid}{OdbconnectResult}(dbObj, ...)

\S4method{dbGetStatement}{OdbconnectResult}(res, ...)

\S4method{dbColumnInfo}{OdbconnectResult}(res, ...)

\S4method{dbGetRowCount}{OdbconnectResult}(res, ...)

\S4method{dbGetRowsAffected}{OdbconnectResult}(res, ...)

\S4method{dbBind}{OdbconnectResult}(res, params, ...)

\S4method{dbReadTable}{OdbconnectConnection,character}(conn, name, ...)
}
\arguments{
\item{object}{Any R object}

\item{drv}{an object that inherits from \code{\linkS4class{DBIDriver}},
or an existing \code{\linkS4class{DBIConnection}}
object (in order to clone an existing connection).}

\item{dsn}{The ODBC dsn to connect to.}

\item{...}{Additional ODBC named arguments to use in the connection, these
will be joined together to form the connection string.}

\item{dbObj}{A object inheriting from \code{\linkS4class{DBIDriver}}
or \code{\linkS4class{DBIConnection}}}

\item{obj}{An R object whose SQL type we want to determine.}

\item{conn}{A \code{\linkS4class{DBIConnection}} object, as produced by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{statement}{a character vector of length 1 containing SQL.}

\item{x}{A character vector to label as being escaped SQL.}

\item{name}{A character string specifying a DBMS table name.}

\item{res}{An object inheriting from \code{\linkS4class{DBIResult}}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{params}{A list of bindings}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package.
}

