% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{dbListTables,OdbcConnection-method}
\alias{dbListTables,OdbcConnection-method}
\alias{dbListFields,OdbcConnection,character-method}
\title{List remote tables and fields for an ODBC connection}
\usage{
\S4method{dbListTables}{OdbcConnection}(
  conn,
  catalog_name = NULL,
  schema_name = NULL,
  table_name = NULL,
  table_type = NULL,
  ...
)

\S4method{dbListFields}{OdbcConnection,character}(
  conn,
  name,
  catalog_name = NULL,
  schema_name = NULL,
  column_name = NULL,
  ...
)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{catalog_name, schema_name, table_name}{Catalog, schema, and table names.

Use \verb{\%} as a wildcard to match any name; use \verb{_} to any any character.}

\item{table_type}{The type of the table to return, the default returns all table types.}

\item{...}{Other parameters passed on to methods.}

\item{column_name}{The name of the column to return, the default returns all columns.}
}
\value{
A character vector of table or field names respectively.
}
\description{
\code{dbListTables()} provides names of remote tables accessible through this
connection; \code{dbListFields()} provides names of columns within a table.
}
