% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\docType{methods}
\name{dbSendQuery,ODBConnectConnection-method}
\alias{dbSendQuery,ODBConnectConnection-method}
\title{Send a query to ODBConnect.}
\usage{
\S4method{dbSendQuery}{ODBConnectConnection}(conn, statement,
  buffer_size = 4096, ...)
}
\description{
Send a query to ODBConnect.
}
\examples{
# This is another good place to put examples
}

