% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{dbUnQuoteIdentifier}
\alias{dbUnQuoteIdentifier}
\alias{dbUnQuoteIdentifier,OdbcConnection,SQL-method}
\alias{dbUnQuoteIdentifier,OdbcConnection,character-method}
\title{Un-Quote identifiers}
\usage{
dbUnQuoteIdentifier(conn, x, ...)

\S4method{dbUnQuoteIdentifier}{OdbcConnection,SQL}(conn, x)

\S4method{dbUnQuoteIdentifier}{OdbcConnection,character}(conn, x)
}
\arguments{
\item{conn}{A subclass of \linkS4class{DBIConnection}, representing
an active connection to an DBMS.}

\item{x}{A character vector to un-quote.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Call this method to generate a string that is unquoted. This is the inverse
of \code{DBI::dbQuoteIdentifier}.
}
