% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_oracle.R, R/hidden.R
\docType{data}
\name{sqlCreateTable}
\alias{sqlCreateTable}
\alias{hidden_aliases}
\title{Internal page for hidden aliases}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
