% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{isTempTable}
\alias{isTempTable}
\title{Helper method used to determine if a table identifier is that
of a temporary table.}
\usage{
isTempTable(conn, name, ...)
}
\arguments{
\item{conn}{OdbcConnection}

\item{name}{Table name}

\item{...}{additional parameters to methods}
}
\description{
Currently implemented only for select back-ends where
we have a use for it (SQL Server, for example).  Generic, in case
we develop a broader use case.
}
