% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-lakebase.R
\docType{class}
\name{lakebase}
\alias{lakebase}
\alias{LakebaseOdbcDriver-class}
\alias{dbConnect,LakebaseOdbcDriver-method}
\title{Helper for Connecting to Databricks Lakebase via ODBC}
\usage{
lakebase()

\S4method{dbConnect}{LakebaseOdbcDriver}(
  drv,
  instance_id,
  workspace = Sys.getenv("DATABRICKS_HOST"),
  database = "databricks_postgres",
  driver = NULL,
  uid = NULL,
  pwd = NULL,
  ...
)
}
\arguments{
\item{drv}{an object that inherits from \link[DBI:DBIDriver-class]{DBI::DBIDriver},
or an existing \link[DBI:DBIConnection-class]{DBI::DBIConnection}
object (in order to clone an existing connection).}

\item{instance_id}{The UUID identifying the Lakebase instance, e.g.
\code{"3860009d-c108-4632-945c-460ba27870e3"}.}

\item{workspace}{The URL of a Databricks workspace, e.g.
\code{"https://adb-1234567890123456.7.azuredatabricks.net"}. Defaults to the
\code{DATABRICKS_HOST} environment variable.}

\item{database}{The database name. Defaults to \code{"databricks_postgres"}.}

\item{driver}{The name of the PostgreSQL ODBC driver, or \code{NULL} to use the
default name.}

\item{uid, pwd}{Manually specify a username and password for authentication.

For Lakebase, use your email as \code{uid} and a Databricks personal access
token as \code{pwd}. Specifying these options will disable automated credential
discovery.}

\item{...}{Further arguments passed on to \code{\link[=dbConnect]{dbConnect()}}.}
}
\value{
An \code{OdbcConnection} object with an active connection to a Databricks
Lakebase instance.
}
\description{
Connect to Databricks Lakebase (managed PostgreSQL service) via the
PostgreSQL ODBC driver.

In particular, the custom \code{dbConnect()} method for the Lakebase ODBC driver
implements a subset of the \href{https://docs.databricks.com/en/dev-tools/auth.html#databricks-client-unified-authentication}{Databricks client unified authentication}
model, with support for personal access tokens, OAuth machine-to-machine
credentials, and OAuth user-to-machine credentials supplied via Posit
Workbench or the Databricks CLI on desktop. It can also detect viewer-based
credentials on Posit Connect if the \pkg{connectcreds} package is
installed. All of these credentials are detected automatically if present
using \href{https://docs.databricks.com/en/dev-tools/auth.html#environment-variables-and-fields-for-client-unified-authentication}{standard environment variables}.
}
\examples{
\dontrun{
DBI::dbConnect(
  odbc::lakebase(),
  instance_id = "3860009d-c108-4632-945c-460ba27870e3"
)

# Use credentials from the viewer (when possible) in a Shiny app
# deployed to Posit Connect.
library(connectcreds)
server <- function(input, output, session) {
  conn <- DBI::dbConnect(
    odbc::lakebase(),
    instance_id = "3860009d-c108-4632-945c-460ba27870e3"
  )
}
}
}
