% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{list_data_sources}
\alias{list_data_sources}
\title{List Available Data Source Names}
\usage{
list_data_sources()
}
\value{
A data frame with two columns.
\describe{
\item{name}{Name of the data source}
\item{description}{Data Source description}
}
}
\description{
List Available Data Source Names
}
