% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\docType{class}
\name{MSSQLConnectionAttributes-class}
\alias{MSSQLConnectionAttributes-class}
\title{MSSQLConnectionAttributes}
\description{
MSSQLConnectionAttributes
}
\details{
A class for MSSQL back-end specific connection attributes
}
\section{Slots}{

\describe{
\item{\code{azure_token}}{Azure Active Directory token, as character string.}
}}

\examples{
\dontrun{
r_attr = new("MSSQLConnectionAttributes", azure_token = .token)
conn <- dbConnect(
  odbc::odbc(),
  dsn = "my_azure_mssql_db",
  Encrypt = "yes",
  attributes = r_attributes)
}
}
