% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{odbcConnectionCatalogs}
\alias{odbcConnectionCatalogs}
\alias{odbcConnectionCatalogs,OdbcConnection-method}
\title{odbcConnectionCatalogs}
\usage{
odbcConnectionCatalogs(conn)

\S4method{odbcConnectionCatalogs}{OdbcConnection}(conn)
}
\arguments{
\item{conn}{OdbcConnection}
}
\description{
This function returns a listing of available
catalogs.
}
