% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{odbcConnectionColumns}
\alias{odbcConnectionColumns}
\alias{odbcConnectionColumns,OdbcConnection,Id-method}
\alias{odbcConnectionColumns,OdbcConnection,character-method}
\title{odbcConnectionColumns}
\usage{
odbcConnectionColumns(conn, name, ...)

\S4method{odbcConnectionColumns}{OdbcConnection,Id}(conn, name,
  column_name = NULL)

\S4method{odbcConnectionColumns}{OdbcConnection,character}(conn, name,
  catalog_name = NULL, schema_name = NULL, column_name = NULL)
}
\arguments{
\item{conn}{OdbcConnection}

\item{name}{table we wish to get information on}

\item{...}{additional parameters to methods}

\item{column_name}{The name of the column to return, the default returns all columns.}

\item{catalog_name}{charaacter catalog where the table is located}

\item{schema_name}{charaacter schema where the table is located}
}
\value{
data.frame with columns
\itemize{
\item{table_cat}
\item{table_schem}
\item{table_name}
\item{column_name}
\item{data_type}
\item{type_name}
\item{column_size}
\item{buffer_length}
\item{decimal_digits}
\item{numeric_precision_radix}
\item{nullable}
\item{column_def}
\item{sql_data_type}
\item{sql_datetime_subtype}
\item{char_octet_length}
\item{ordinal_position}
}
}
\description{
For a given table this function recturns detailed information on
all fields / columns.  The expectation is that this is a relatively thin
wrapper around the ODBC SQLColumns function call.
}
\details{
For more information see SQLColumns ODBC function
}
\seealso{
The ODBC documentation on \href{https://docs.microsoft.com/en-us/sql/odbc/reference/syntax/sqlcolumns-function}{SQLColumns}
for further details on the expected output.
}
