% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{odbcConnectionTableTypes}
\alias{odbcConnectionTableTypes}
\alias{odbcConnectionTableTypes,OdbcConnection-method}
\title{odbcConnectionTableTypes}
\usage{
odbcConnectionTableTypes(conn)

\S4method{odbcConnectionTableTypes}{OdbcConnection}(conn)
}
\arguments{
\item{conn}{OdbcConnection}
}
\description{
This function returns a listing of table
types available in database.
}
