% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-odbc-data-type.R
\name{odbcDataType}
\alias{odbcDataType}
\title{Return the corresponding ODBC data type for an R object}
\usage{
odbcDataType(con, obj, ...)
}
\arguments{
\item{con}{A driver connection object, as returned by \code{dbConnect()}.}

\item{obj}{An R object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Corresponding SQL type for the \code{obj}.
}
\description{
This is used when creating a new table with \code{dbWriteTable()}.
Databases with default methods defined are:
\itemize{
\item MySQL
\item PostgreSQL
\item SQL Server
\item Oracle
\item SQLite
\item Spark
\item Hive
\item Impala
\item Redshift
\item Vertica
\item BigQuery
\item Teradata
\item Access
\item Snowflake
}
}
\details{
If you are using a different database and \code{dbWriteTable()} fails with a SQL
parsing error the default method is not appropriate, you will need to write
a new method. The object type for your method will be the database name
retrieved by \code{dbGetInfo(con)$dbms.name}. Use the documentation provided with
your database to determine appropriate values for each R data type.
}
