% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\docType{class}
\name{OdbcDriver}
\alias{OdbcDriver}
\alias{OdbcDriver-class}
\alias{dbDataType,OdbcDriver,ANY-method}
\alias{dbDataType,OdbcDriver,list-method}
\alias{dbGetInfo,OdbcDriver-method}
\alias{dbIsValid,OdbcDriver-method}
\alias{show,OdbcDriver-method}
\title{Odbc Driver Methods}
\usage{
\S4method{show}{OdbcDriver}(object)

\S4method{dbDataType}{OdbcDriver,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{OdbcDriver,list}(dbObj, obj, ...)

\S4method{dbIsValid}{OdbcDriver}(dbObj, ...)

\S4method{dbGetInfo}{OdbcDriver}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{A object inheriting from \code{\linkS4class{DBIDriver}}
or \code{\linkS4class{DBIConnection}}}

\item{obj}{An R object whose SQL type we want to determine.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for OdbcDriver objects.
}

