% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{odbcListColumns}
\alias{odbcListColumns}
\title{List columns in an object.}
\usage{
odbcListColumns(connection, ...)
}
\arguments{
\item{connection}{A connection object, as returned by \code{dbConnect()}.}

\item{...}{Parameters specifying the object.}
}
\value{
A data frame with \code{name} and \code{type} columns, listing the
object's fields.
}
\description{
Lists the names and types of each column (field) of a specified object.
}
\details{
The object to inspect must be specified as one of the arguments
(e.g. \code{table = "employees"}); depending on the driver and underlying
data store, additional specification arguments may be required.
}
