% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{class}
\name{OdbconnectConnection}
\alias{OdbconnectConnection}
\alias{OdbconnectConnection-class}
\alias{dbBegin,OdbconnectConnection-method}
\alias{dbCommit,OdbconnectConnection-method}
\alias{dbDataType,OdbconnectConnection,ANY-method}
\alias{dbDisconnect,OdbconnectConnection-method}
\alias{dbExistsTable,OdbconnectConnection,character-method}
\alias{dbGetInfo,OdbconnectConnection-method}
\alias{dbGetQuery,OdbconnectConnection,character-method}
\alias{dbIsValid,OdbconnectConnection-method}
\alias{dbListFields,OdbconnectConnection,character-method}
\alias{dbListTables,OdbconnectConnection-method}
\alias{dbQuoteIdentifier,OdbconnectConnection,character-method}
\alias{dbQuoteString,OdbconnectConnection,character-method}
\alias{dbRemoveTable,OdbconnectConnection,character-method}
\alias{dbRollback,OdbconnectConnection-method}
\alias{dbSendQuery,OdbconnectConnection,character-method}
\alias{dbSendStatement,OdbconnectConnection,character-method}
\alias{show,OdbconnectConnection-method}
\title{Odbconnect Connection Methods}
\usage{
\S4method{show}{OdbconnectConnection}(object)

\S4method{dbIsValid}{OdbconnectConnection}(dbObj, ...)

\S4method{dbDisconnect}{OdbconnectConnection}(conn, ...)

\S4method{dbSendQuery}{OdbconnectConnection,character}(conn, statement, ...)

\S4method{dbSendStatement}{OdbconnectConnection,character}(conn, statement, ...)

\S4method{dbDataType}{OdbconnectConnection,ANY}(dbObj, obj, ...)

\S4method{dbQuoteString}{OdbconnectConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{OdbconnectConnection,character}(conn, x, ...)

\S4method{dbListTables}{OdbconnectConnection}(conn, ...)

\S4method{dbExistsTable}{OdbconnectConnection,character}(conn, name, ...)

\S4method{dbListFields}{OdbconnectConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{OdbconnectConnection,character}(conn, name, ...)

\S4method{dbGetInfo}{OdbconnectConnection}(dbObj, ...)

\S4method{dbGetQuery}{OdbconnectConnection,character}(conn, statement, ...)

\S4method{dbBegin}{OdbconnectConnection}(conn, ...)

\S4method{dbCommit}{OdbconnectConnection}(conn, ...)

\S4method{dbRollback}{OdbconnectConnection}(conn, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{An object inheriting from \code{\linkS4class{DBIObject}},
i.e. \code{\linkS4class{DBIDriver}}, \code{\linkS4class{DBIConnection}},
or a \code{\linkS4class{DBIResult}}}

\item{...}{Other arguments to methods.}

\item{conn}{A \code{\linkS4class{DBIConnection}} object, as produced by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{statement}{a character vector of length 1 containing SQL.}

\item{obj}{An R object whose SQL type we want to determine.}

\item{x}{A character vector to label as being escaped SQL.}

\item{name}{A character string specifying a DBMS table name.}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for OdbconnectConnection objects.
}

