% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R, R/Connection.R
\docType{class}
\name{ODBConnectDriver-class}
\alias{ODBConnectDriver-class}
\alias{dbConnect,ODBConnectDriver-method}
\alias{dbGetInfo,ODBConnectDriver-method}
\alias{dbUnloadDriver,ODBConnectDriver-method}
\alias{show,ODBConnectDriver-method}
\title{Driver for an ODBC database.}
\usage{
\S4method{dbUnloadDriver}{ODBConnectDriver}(drv, ...)

\S4method{show}{ODBConnectDriver}(object)

\S4method{dbGetInfo}{ODBConnectDriver}(dbObj, what = "", ...)

\S4method{dbConnect}{ODBConnectDriver}(drv, connection_string, ...)
}
\arguments{
\item{drv}{An object created by \code{ODBC()}}
}
\description{
Driver for an ODBC database.
}
\examples{
\dontrun{
db <- dbConnect(odbconnect::odbconnect(), "DSN=database1")
dbWriteTable(db, "mtcars", mtcars)
dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == 4")
}
}
\keyword{internal}

