% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\docType{class}
\name{OdbconnectDriver}
\alias{OdbconnectDriver}
\alias{OdbconnectDriver-class}
\alias{dbDataType,OdbconnectDriver,ANY-method}
\alias{dbDataType,OdbconnectDriver,list-method}
\alias{dbGetInfo,OdbconnectDriver-method}
\alias{dbIsValid,OdbconnectDriver-method}
\alias{show,OdbconnectDriver-method}
\title{Odbconnect Driver Methods}
\usage{
\S4method{show}{OdbconnectDriver}(object)

\S4method{dbDataType}{OdbconnectDriver,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{OdbconnectDriver,list}(dbObj, obj, ...)

\S4method{dbIsValid}{OdbconnectDriver}(dbObj, ...)

\S4method{dbGetInfo}{OdbconnectDriver}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{A object inheriting from \code{\linkS4class{DBIDriver}}
or \code{\linkS4class{DBIConnection}}}

\item{obj}{An R object whose SQL type we want to determine.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for OdbconnectDriver objects.
}

