% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\docType{class}
\name{OdbconnectResult}
\alias{OdbconnectResult}
\alias{OdbconnectResult-class}
\alias{dbBind,OdbconnectResult-method}
\alias{dbClearResult,OdbconnectResult-method}
\alias{dbColumnInfo,OdbconnectResult-method}
\alias{dbFetch,OdbconnectResult-method}
\alias{dbGetInfo,OdbconnectResult-method}
\alias{dbGetRowCount,OdbconnectResult-method}
\alias{dbGetRowsAffected,OdbconnectResult-method}
\alias{dbGetStatement,OdbconnectResult-method}
\alias{dbHasCompleted,OdbconnectResult-method}
\alias{dbIsValid,OdbconnectResult-method}
\title{Odbconnect Result Methods}
\usage{
\S4method{dbClearResult}{OdbconnectResult}(res, ...)

\S4method{dbFetch}{OdbconnectResult}(res, n = -1, ..., row.names = NA)

\S4method{dbHasCompleted}{OdbconnectResult}(res, ...)

\S4method{dbGetInfo}{OdbconnectResult}(dbObj, ...)

\S4method{dbIsValid}{OdbconnectResult}(dbObj, ...)

\S4method{dbGetStatement}{OdbconnectResult}(res, ...)

\S4method{dbColumnInfo}{OdbconnectResult}(res, ...)

\S4method{dbGetRowCount}{OdbconnectResult}(res, ...)

\S4method{dbGetRowsAffected}{OdbconnectResult}(res, ...)

\S4method{dbBind}{OdbconnectResult}(res, params, ...)
}
\arguments{
\item{res}{An object inheriting from \code{\linkS4class{DBIResult}}.}

\item{...}{Other arguments passed on to methods.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{dbObj}{An object inheriting from \code{\linkS4class{DBIObject}},
i.e. \code{\linkS4class{DBIDriver}}, \code{\linkS4class{DBIConnection}},
or a \code{\linkS4class{DBIResult}}}

\item{params}{A list of bindings}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for OdbconnectResult objects.
}

