% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{odbcPreviewQuery}
\alias{odbcPreviewQuery}
\alias{odbcPreviewQuery.OdbcConnection}
\alias{odbcPreviewQuery.Microsoft SQL Server}
\alias{odbcPreviewQuery.Oracle}
\title{Create a preview query.}
\usage{
odbcPreviewQuery(connection, rowLimit, name)

\method{odbcPreviewQuery}{OdbcConnection}(connection, rowLimit, name)

\method{odbcPreviewQuery}{`Microsoft SQL Server`}(connection, rowLimit, name)

\method{odbcPreviewQuery}{Oracle}(connection, rowLimit, name)
}
\arguments{
\item{connection}{A connection object, as returned by \code{dbConnect()}.}

\item{rowLimit}{The maximum number of rows to display.}

\item{name}{Name of the object to be previewed}
}
\description{
Optimize against the rowLimit argument.  S3 since some
back-ends do not parse the LIMIT syntax.  Internal, not expected that
users would interact with this method.
}
