% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-result-arrow.R
\docType{methods}
\name{OdbcResultArrow}
\alias{OdbcResultArrow}
\alias{OdbcResultArrow-class}
\alias{show,OdbcResultArrow-method}
\alias{dbClearResult,OdbcResultArrow-method}
\alias{dbIsValid,OdbcResultArrow-method}
\alias{dbHasCompleted,OdbcResultArrow-method}
\alias{dbGetStatement,OdbcResultArrow-method}
\alias{dbColumnInfo,OdbcResultArrow-method}
\alias{dbGetRowCount,OdbcResultArrow-method}
\alias{dbGetRowsAffected,OdbcResultArrow-method}
\alias{dbGetInfo,OdbcResultArrow-method}
\alias{dbBind,OdbcResultArrow-method}
\title{Odbc Arrow result methods}
\usage{
\S4method{show}{OdbcResultArrow}(object)

\S4method{dbClearResult}{OdbcResultArrow}(res, ...)

\S4method{dbIsValid}{OdbcResultArrow}(dbObj, ...)

\S4method{dbHasCompleted}{OdbcResultArrow}(res, ...)

\S4method{dbGetStatement}{OdbcResultArrow}(res, ...)

\S4method{dbColumnInfo}{OdbcResultArrow}(res, ...)

\S4method{dbGetRowCount}{OdbcResultArrow}(res, ...)

\S4method{dbGetRowsAffected}{OdbcResultArrow}(res, ...)

\S4method{dbGetInfo}{OdbcResultArrow}(dbObj, ...)

\S4method{dbBind}{OdbcResultArrow}(res, params, ...)
}
\arguments{
\item{res, dbObj, object}{An \code{OdbcResultArrow} object.}

\item{...}{Passed on to the corresponding \link{OdbcResult} method.}

\item{params}{Query parameters, see \code{\link[DBI:dbBind]{DBI::dbBind()}}.}
}
\description{
\code{OdbcResultArrow} objects are returned by \code{\link[DBI:dbSendQueryArrow]{DBI::dbSendQueryArrow()}}. They
wrap an \link{OdbcResult} and implement the methods defined in the \code{DBI} package
for \link[DBI:DBIResultArrow-class]{DBI::DBIResultArrow} objects by delegating to it. See
\link{DBI-arrow} for the native Arrow interface of odbc.
}
\keyword{internal}
