% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{odbcSetBigIntMapping}
\alias{odbcSetBigIntMapping}
\title{Set the BIGINT type mapping}
\usage{
odbcSetBigIntMapping(conn, bigint = c("integer64", "integer", "numeric",
  "character"))
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{bigint}{An R type that \code{SQL BIGINT} type should be mapped to while
reading from the database.}
}
\description{
Set the BIGINT type mapping
}
