% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{odbcSetMetadataId}
\alias{odbcSetMetadataId}
\title{Set METADATA_ID}
\usage{
odbcSetMetadataId(conn, val)
}
\description{
Configure whether the string argument in catalog functions is treated as an
identifier (\code{TRUE}) or a search pattern (\code{FALSE}).  The default
is \code{FALSE}.
}
\examples{
\dontrun{
odbcSetMetadataId(conn, TRUE)
}

}
