% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{odbcSetTransactionIsolationLevel}
\alias{odbcSetTransactionIsolationLevel}
\title{Set the Transaction Isolation Level for a Connection}
\usage{
odbcSetTransactionIsolationLevel(conn, levels)
}
\arguments{
\item{levels}{One or more of \Sexpr\link{stage=render, results=rd}{odbc:::choices_rd(names(odbc:::transactionLevels()))}.}
}
\description{
Set the Transaction Isolation Level for a Connection
}
