% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc-connection.R
\name{quote_value}
\alias{quote_value}
\title{Quote special character when connecting}
\usage{
quote_value(x)
}
\arguments{
\item{x}{A string to quote.}
}
\value{
A quoted string, wrapped in \code{I()}.
}
\description{
When connecting to a database using odbc, all the arguments are concatenated
into a single connection string that looks like \verb{name1=value1;name2=value2}.
That means if your value contains \code{=} or \verb{;} then it needs to be quoted.
Other rules mean that it's generally a good idea to quote any text that
contains \verb{\{}, \verb{\}} or white space.

This function attempts to quote a string in a way that should hopefully
work for most drivers. Unfortunately, there doesn't seem to be an approach
that works everywhere, but hopefully this function solves most common
issues.
}
\examples{
quote_value("abc")
quote_value("ab'c")

# Real usage is more likely to look like:
\dontrun{
library(DBI)

con <- dbConnect(
  odbc::odbc(),
  dsn = "reallycooldatabase"
  password = odbc::quote_value(Sys.getenv("MY_PASSWORD"))
)
}
}
