% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{setBigintMapping}
\alias{setBigintMapping}
\title{Set the BIGINT type mapping}
\usage{
setBigintMapping(map.to = c("integer64", "integer", "numeric"))
}
\arguments{
\item{map.to}{the R type (as string) that SQL BIGINT type should be mapped to.}
}
\description{
Set the BIGINT type mapping
}
