% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{SQLTable}
\alias{SQLTable}
\title{SQL Table}
\usage{
SQLTable(con, name, schema = character(), catalog = character())
}
\arguments{
\item{con}{A \code{DBIConnection} object.}

\item{name}{A character vector with the table name.}

\item{schema}{A character vector with the table schema (optional).}

\item{catalog}{A character vector with the table catalog (optional).}
}
\description{
A class to refer to a specific SQL table, including those in non-default
catalogs and schemas.
}
