% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\docType{class}
\name{ConnectionAttributes-class}
\alias{ConnectionAttributes-class}
\title{ConnectionAttributes}
\description{
ConnectionAttributes
}
\details{
A class that all back-end-specific connection attribute classes
are derived from.  If there is an attribute that is shared across back-ends
it belongs here.
}
