% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R, R/driver-db2.R, R/driver-hana.R,
%   R/driver-hive.R, R/driver-oracle.R, R/driver-spark.R, R/driver-terdata.R
\docType{class}
\name{DBI-classes}
\alias{DBI-classes}
\alias{DB2/AIX64-class}
\alias{HDB-class}
\alias{Hive-class}
\alias{Oracle-class}
\alias{Spark SQL-class}
\alias{Teradata-class}
\title{DBI classes}
\description{
DBI classes
}
\keyword{internal}
