% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R, R/driver-access.R, R/driver-bigquery.R,
%   R/driver-db2.R, R/driver-hana.R, R/driver-hive.R, R/driver-impala.R,
%   R/driver-mysql.R, R/driver-netezza.R, R/driver-postgres.R,
%   R/driver-redshift.R, R/driver-snowflake.R, R/driver-spark.R,
%   R/driver-sqlite.R, R/driver-teradata.R, R/driver-vertica.R
\docType{class}
\name{DBI-classes}
\alias{DBI-classes}
\alias{ACCESS-class}
\alias{BigQuery-class}
\alias{DB2/AIX64-class}
\alias{HDB-class}
\alias{Hive-class}
\alias{Impala-class}
\alias{MySQL-class}
\alias{NetezzaSQL-class}
\alias{PostgreSQL-class}
\alias{Redshift-class}
\alias{Snowflake-class}
\alias{Spark SQL-class}
\alias{SQLite-class}
\alias{Teradata-class}
\alias{Vertica Database-class}
\title{DBI classes}
\description{
DBI classes
}
\keyword{internal}
