% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-table.R, R/dbi.R, R/driver-db2.R,
%   R/driver-hana.R, R/driver-hive.R, R/driver-teradata.R
\name{sqlData,OdbcConnection-method}
\alias{sqlData,OdbcConnection-method}
\alias{DBI-methods}
\alias{sqlCreateTable,DB2/AIX64-method}
\alias{sqlCreateTable,HDB-method}
\alias{dbQuoteString,Hive,character-method}
\alias{sqlCreateTable,Teradata-method}
\title{Unimportant DBI methods}
\usage{
\S4method{sqlData}{OdbcConnection}(con, value, row.names = NA, ...)

\S4method{sqlCreateTable}{HDB}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)

\S4method{dbQuoteString}{Hive,character}(conn, x, ...)

\S4method{sqlCreateTable}{Teradata}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)
}
\arguments{
\item{...}{Other parameters passed on to methods.}

\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}
}
\description{
Unimportant DBI methods
}
\details{
If the \code{value} argument has an attribute \code{.odbc.transformed} that
is set to TRUE, then the method immediately returns the input value without
modifications.  This is an optimization to avoid multiple methods that are
chained together from needlessly iterating across the columns of
the data.frame.  It is unlikely that users would ever need to set ( or
interogate ) this attribute directly.
}
\keyword{internal}
