% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-table.R, R/dbi.R, R/driver-db2.R,
%   R/driver-hana.R, R/driver-hive.R, R/driver-teradata.R
\name{sqlData,OdbcConnection-method}
\alias{sqlData,OdbcConnection-method}
\alias{DBI-methods}
\alias{sqlCreateTable,DB2/AIX64-method}
\alias{odbcConnectionTables,DB2/AIX64,character-method}
\alias{sqlCreateTable,HDB-method}
\alias{dbQuoteString,Hive,character-method}
\alias{sqlCreateTable,Teradata-method}
\title{Unimportant DBI methods}
\usage{
\S4method{sqlData}{OdbcConnection}(con, value, row.names = NA, ...)

\S4method{sqlCreateTable}{HDB}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)

\S4method{dbQuoteString}{Hive,character}(conn, x, ...)

\S4method{sqlCreateTable}{Teradata}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)
}
\arguments{
\item{...}{Other parameters passed on to methods.}

\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{name}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}
}
\description{
Unimportant DBI methods
}
\keyword{internal}
