% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-table.R, R/dbi.R, R/driver-db2.R,
%   R/driver-hana.R, R/driver-hive.R, R/driver-teradata.R
\name{sqlData,OdbcConnection-method}
\alias{sqlData,OdbcConnection-method}
\alias{DBI-methods}
\alias{sqlCreateTable,DB2/AIX64-method}
\alias{odbcConnectionTables,DB2/AIX64,character-method}
\alias{sqlCreateTable,HDB-method}
\alias{dbQuoteString,Hive,character-method}
\alias{sqlCreateTable,Teradata-method}
\title{Unimportant DBI methods}
\usage{
\S4method{sqlData}{OdbcConnection}(con, value, row.names = NA, ...)

\S4method{sqlCreateTable}{HDB}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)

\S4method{dbQuoteString}{Hive,character}(conn, x, ...)

\S4method{sqlCreateTable}{Teradata}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)
}
\description{
Unimportant DBI methods
}
\keyword{internal}
