% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_oracle.R, R/hidden.R
\docType{methods}
\name{sqlCreateTable,Oracle-method}
\alias{sqlCreateTable,Oracle-method}
\alias{hidden_aliases}
\title{Internal page for hidden aliases}
\usage{
\S4method{sqlCreateTable}{Oracle}(con, table, fields, row.names = NA,
  temporary = FALSE, ...)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
