% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R, R/hidden.R
\docType{data}
\name{sqlCreateTable,Oracle-method}
\alias{sqlCreateTable,Oracle-method}
\alias{sqlCreateTable,Teradata-method}
\alias{dbListTables,Teradata-method}
\alias{sqlCreateTable,HDB-method}
\alias{hidden_aliases}
\title{Internal page for hidden aliases}
\usage{
\S4method{sqlCreateTable}{Oracle}(
  con,
  table,
  fields,
  field.types = NULL,
  row.names = NA,
  temporary = FALSE,
  ...
)

\S4method{sqlCreateTable}{Teradata}(
  con,
  table,
  fields,
  field.types = NULL,
  row.names = NA,
  temporary = FALSE,
  ...
)

\S4method{dbListTables}{Teradata}(conn, ...)

\S4method{sqlCreateTable}{HDB}(
  con,
  table,
  fields,
  field.types = NULL,
  row.names = NA,
  temporary = FALSE,
  ...
)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
