% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc-package.R
\docType{package}
\name{odbc-package}
\alias{odbc-package}
\title{odbc: Connect to ODBC Compatible Databases (using the DBI Interface)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A DBI-compatible interface to ODBC databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://odbc.r-dbi.org}
  \item \url{https://github.com/r-dbi/odbc}
  \item \url{https://solutions.posit.co/connections/db/}
  \item Report bugs at \url{https://github.com/r-dbi/odbc/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
  \item Jim Hester
  \item Oliver Gjoneski
  \item Simon Couch \email{simon.couch@posit.co}
}

Other contributors:
\itemize{
  \item lexicalunit (nanodbc library) [copyright holder]
  \item Google Inc. (cctz library) [copyright holder]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
