% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\docType{package}
\name{odbc-package}
\title{odbc: Connect to ODBC Compatible Databases (using the DBI
    Interface)}
\description{
A DBI-compatible interface to ODBC databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-dbi/odbc}
  \item \url{https://db.rstudio.com}
  \item Report bugs at \url{https://github.com/r-dbi/odbc/issues}
}

}
\author{
\strong{Maintainer}: Jim Hester \email{jim.hester@rstudio.com}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item Oliver Gjoneski (detule) [contributor]
  \item lexicalunit (nanodbc library) [copyright holder]
  \item Google Inc. (cctz library) [copyright holder]
  \item RStudio [copyright holder, funder]
}

}
