% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{odbcConnectionColumns}
\alias{odbcConnectionColumns}
\title{odbcConnectionColumns}
\usage{
odbcConnectionColumns(conn, name, ...)
}
\arguments{
\item{conn}{OdbcConnection}
}
\value{
data.frame with columns
\itemize{
\item{table_cat}
\item{table_schem}
\item{table_name}
\item{column_name}
\item{data_type}
\item{type_name}
\item{column_size}
\item{buffer_length}
\item{decimal_digits}
\item{numeric_precision_radix}
\item{nullable}
\item{column_def}
\item{sql_data_type}
\item{sql_datetime_subtype}
\item{char_octet_length}
\item{ordinal_position}
}
}
\description{
For a given table this function recturns detailed information on
all fields / columns.  The expectation is that this is a relatively thin
wrapper around the ODBC SQLColumns function call.
}
\details{
For more information see SQLColumns ODBC function
}
