% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{odbcConnectionIcon}
\alias{odbcConnectionIcon}
\title{Get an icon representing a connection.}
\usage{
odbcConnectionIcon(connection)
}
\arguments{
\item{connection}{A connection object, as returned by \code{dbConnect()}.}
}
\value{
The path to an icon file on disk.
}
\description{
Return the path on disk to an icon representing a connection.
}
\details{
The icon returned should be a 32x32 square image file.
}
\keyword{internal}
