% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi-driver.R
\docType{class}
\name{OdbcDriver}
\alias{OdbcDriver}
\alias{OdbcDriver-class}
\alias{show,OdbcDriver-method}
\alias{dbDataType,OdbcDriver,ANY-method}
\alias{dbDataType,OdbcDriver,list-method}
\alias{dbDataType,OdbcDriver,data.frame-method}
\alias{dbIsValid,OdbcDriver-method}
\alias{dbGetInfo,OdbcDriver-method}
\title{Odbc Driver Methods}
\usage{
\S4method{show}{OdbcDriver}(object)

\S4method{dbDataType}{OdbcDriver,ANY}(dbObj, obj, ...)

\S4method{dbDataType}{OdbcDriver,list}(dbObj, obj, ...)

\S4method{dbDataType}{OdbcDriver,data.frame}(dbObj, obj, ...)

\S4method{dbIsValid}{OdbcDriver}(dbObj, ...)

\S4method{dbGetInfo}{OdbcDriver}(dbObj, ...)
}
\arguments{
\item{object}{Any R object}

\item{dbObj}{A object inheriting from \link[DBI:DBIDriver-class]{DBI::DBIDriver}
or \link[DBI:DBIConnection-class]{DBI::DBIConnection}}

\item{obj}{An R object whose SQL type we want to determine.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for OdbcDriver objects.
}
\keyword{internal}
