% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{odbcListDataSources}
\alias{odbcListDataSources}
\title{List Available Data Source Names}
\usage{
odbcListDataSources()
}
\value{
A data frame with two columns.
\describe{
\item{name}{Name of the data source}
\item{description}{Data Source description}
}
}
\description{
List the available data sources on your system. See the \href{https://github.com/r-dbi/odbc#dsn-configuration-files}{DSN Configuration files} section of the
package README for details on how to install data sources for the most
common databases.
}
