% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\docType{methods}
\name{OdbcResult}
\alias{OdbcResult}
\alias{OdbcResult-class}
\alias{dbClearResult,OdbcResult-method}
\alias{dbFetch,OdbcResult-method}
\alias{dbHasCompleted,OdbcResult-method}
\alias{dbGetInfo,OdbcResult-method}
\alias{dbIsValid,OdbcResult-method}
\alias{dbGetStatement,OdbcResult-method}
\alias{dbColumnInfo,OdbcResult-method}
\alias{dbGetRowCount,OdbcResult-method}
\alias{dbGetRowsAffected,OdbcResult-method}
\alias{dbBind,OdbcResult-method}
\title{Odbc Result Methods}
\usage{
\S4method{dbClearResult}{OdbcResult}(res, ...)

\S4method{dbFetch}{OdbcResult}(res, n = -1, ..., row.names = NA)

\S4method{dbHasCompleted}{OdbcResult}(res, ...)

\S4method{dbGetInfo}{OdbcResult}(dbObj, ...)

\S4method{dbIsValid}{OdbcResult}(dbObj, ...)

\S4method{dbGetStatement}{OdbcResult}(res, ...)

\S4method{dbColumnInfo}{OdbcResult}(res, ...)

\S4method{dbGetRowCount}{OdbcResult}(res, ...)

\S4method{dbGetRowsAffected}{OdbcResult}(res, ...)

\S4method{dbBind}{OdbcResult}(res, params, ...)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{...}{Other arguments passed on to methods.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{params}{A list of bindings, named or unnamed.}
}
\description{
Implementations of pure virtual functions defined in the \code{DBI} package
for OdbcResult objects.
}
