% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-oracle.R
\docType{class}
\name{Oracle-class}
\alias{Oracle-class}
\alias{sqlCreateTable,Oracle-method}
\alias{odbcConnectionTables,Oracle,character-method}
\alias{odbcConnectionColumns_,Oracle,character-method}
\title{Oracle}
\usage{
\S4method{sqlCreateTable}{Oracle}(
  con,
  table,
  fields,
  row.names = NA,
  temporary = FALSE,
  ...,
  field.types = NULL
)

\S4method{odbcConnectionTables}{Oracle,character}(
  conn,
  name,
  catalog_name = NULL,
  schema_name = NULL,
  table_type = NULL,
  exact = FALSE
)
}
\description{
Details of Oracle methods for odbc and DBI generics.

\subsection{\code{odbcConnectionTables()}}{

Method for an internal function that otherwise relies on the \code{SQLTables}
ODBC API. While this method is much faster than the OEM implementation, it
does not look through synonyms.
}

\subsection{\code{odbcConnectionColumns()}}{

Query, rather than use \code{SQLColumns} ODBC API, since we bind a \code{BIGINT} to
one of the column results and Oracle's OEM driver can't handle it.
}
}
\keyword{internal}
