% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\name{setBigintMapping}
\alias{setBigintMapping}
\title{Set the BIGINT type mapping}
\usage{
setBigintMapping(conn, bigint = c("integer64", "integer", "numeric"))
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{bigint}{An R type name that SQL BIGINT type should be mapped to while
reading from the database.}
}
\description{
Set the BIGINT type mapping
}
