% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver-snowflake.R
\name{odbcConnectionColumns_,Snowflake,character-method}
\alias{odbcConnectionColumns_,Snowflake,character-method}
\alias{dbExistsTableForWrite,Snowflake,character-method}
\title{If the catalog, or the schema arguments are NULL, attempt to infer
by querying for CURRENT_DATABASE() and CURRENT_SCHEMA().  We do this
to aid with performance, as the SQLTables method is more performant
when restricted to a particular DB/schema.}
\usage{
\S4method{dbExistsTableForWrite}{Snowflake,character}(conn, name, ...)
}
\description{
\subsection{\code{odbcConnectionColumns()}}{

If the catalog, or the schema arguments are NULL, attempt to infer
by querying for CURRENT_DATABASE() and CURRENT_SCHEMA().  We do this
to aid with performance, as the SQLColumns method is more performant
when restricted to a particular DB/schema.
}
}
