% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{isTempTable,Microsoft SQL Server,character-method}
\alias{isTempTable,Microsoft SQL Server,character-method}
\alias{sqlCreateTable,Microsoft SQL Server-method}
\title{SQL Server specific implementation.}
\description{
Local temp tables are stored as
\code{ [tempdb].[dbo].[#name]________(padding using underscores)[numeric identifier] }

Will warn user if \code{temporary} is set to TRUE but table name does not conform
to local temp table naming conventions.  If writing to a global temp table, user
should not set the temporary flag to TRUE.
}
\details{
True if:
\itemize{
\item If catalog_name is supplied it must equal "temdb" or "\%" ( wildcard )
\item Name must start with "#" followd by a non-"#" character
}

In both cases a simple CREATE TABLE statement is used / the table identifier is
is the differentiator ( viz-a-viz creating a non-temp table ).
}
